Move the SiteMap into a new controller and have it start to generate an index.
